feat[st]: basic fulfillment strategy as choice for challenge resolving
