+++ SmartTutor/ContentModel/LearningObjects/ChallengeModel/FulfillmentStrategy/BasicChallengeFulfillment.cs
@@ -1,6 +1,7 @@
 ﻿using RepositoryCompiler.CodeModel.CaDETModel.CodeItems;
-using System;
+using SmartTutor.ContentModel.LearningObjects.ChallengeModel.MetricRules;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace SmartTutor.ContentModel.LearningObjects.ChallengeModel.FulfillmentStrategy
 {
@@ -8,7 +9,31 @@ namespace SmartTutor.ContentModel.LearningObjects.ChallengeModel.FulfillmentStra
     {
         public override bool CheckChallengeFulfillment(List<CaDETClass> submittetClasses, Challenge challenge)
         {
-            throw new NotImplementedException();
+            List<CaDETMember> submittedMethods = GetMethodsFromClasses(submittetClasses);
+            return ValidateClassMetricRules(submittetClasses, challenge) && ValidateMethodMetricRules(submittedMethods, challenge);
+        }
+
+        private List<CaDETMember> GetMethodsFromClasses(List<CaDETClass> caDETClasses)
+        {
+            return caDETClasses.SelectMany(c => c.Members.Where(m => m.Type.Equals(CaDETMemberType.Method))).ToList();
+        }
+
+        private bool ValidateClassMetricRules(List<CaDETClass> caDETClasses, Challenge challenge)
+        {
+            foreach (CaDETClass caDETClass in caDETClasses)
+                foreach (MetricRangeRule classMetricRule in challenge.ClassMetricRules)
+                    if (!classMetricRule.MetricMeetsRequirements(caDETClass.Metrics))
+                        return false;
+            return true;
+        }
+
+        private bool ValidateMethodMetricRules(List<CaDETMember> caDETMethods, Challenge challenge)
+        {
+            foreach (CaDETMember caDETMethod in caDETMethods)
+                foreach (MetricRangeRule methodMetricRule in challenge.MethodMetricRules)
+                    if (!methodMetricRule.MetricMeetsRequirements(caDETMethod.Metrics))
+                        return false;
+            return true;
         }
     }
 }
